Add main script
Add script to resize the terminal window based on the number of files
desired open in vim side by side.  Currently it is a shell script.  I
think I might want to change to Python so that I can do the math and
input checking in a smoother way.  In Python, we just need to print the
escape sequence with the print function.  Man, that was redundant...